refactor: use one class for tests
